requirements.txt
Pillow==8.3.1
python-dotenv==0.18.0